restructure: move tableoutput to be an internal package
