+++ internal/tableoutput/table.go
